requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
SQLAlchemy==2.0.34
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
